+++ verification/verification.R
@@ -8,7 +8,7 @@
 #
 # Arseniy Khvorov
 # Created 2019/10/17
-# Last edit 2019/10/17
+# Last edit 2019/10/21
 
 library(sclr)
 library(broom)
@@ -21,51 +21,41 @@ plan(multiprocess)
 
 this_folder <- "verification"
 
-simulate_ideal_data <- function(n_sample = 1e4, 
-                                lambda = 0.5, beta_0 = -5, beta_logtitre = 2,
-                                logtitre_mean = 2, logtitre_sd = 2,
-                                seed = NULL) {
-  if (!is.null(seed)) set.seed(seed)
-  ideal_data <- tibble(.rows = n_sample) %>%
-    mutate(
-      logtitre = rnorm(n(), logtitre_mean, logtitre_sd),
-      status = rbinom(
-        n(), 1, lambda / (1 + exp(beta_0 + beta_logtitre * logtitre))
-      )
-    )
-  attr(ideal_data, "true_values") <- tibble(
-    term = c("lambda", "beta_0", "beta_logtitre"),
-    true_value = c(lambda, beta_0, beta_logtitre)
+simulate_ideal_data <- function(n, lambda, beta_0, covariate_list, seed) {
+  sclr_ideal_data(
+    n, lambda, beta_0, covariate_list, 
+    outcome_name = "status",
+    seed = seed,
+    attach_true_vals = TRUE,
+    attach_seed = TRUE
   )
-  attr(ideal_data, "seed") <- seed
-  ideal_data
 }
 
-fit_sclr_model <- function(data, formula = status ~ logtitre) {
-  tidy_fit <- sclr(formula, data) %>% tidy()
+fit_sclr_model <- function(data, covariate_list) {
+  formula <- paste0("status~", paste(names(covariate_list), collapse = "+"))
+  tidy_fit <- sclr(as.formula(formula), data) %>% tidy()
   if (!is.null(attr(data, "seed"))) tidy_fit$seed <- attr(data, "seed")
   inner_join(tidy_fit, attr(data, "true_values"), by = "term")
 }
 
-simulate_one <- function(index = NULL, 
-                           init_seed = NULL, 
-                           formula = status ~ logtitre, 
-                           ...) {
-  if (is.null(init_seed)) seed <- NULL
-  else if (is.null(index)) seed <- init_seed
-  else seed <- init_seed + index
+simulate_one <- function(index = NULL, init_seed = NULL, ...) {
+  if (is.null(init_seed)) {
+    seed <- NULL
+  } else if (is.null(index)) {
+    seed <- init_seed
+  } else {
+    seed <- init_seed + index
+  }
   args <- list(...)
   args$seed <- seed
-  result <- do.call(simulate_ideal_data, args) %>% fit_sclr_model(formula)
+  result <- do.call(simulate_ideal_data, args) %>%
+    fit_sclr_model(args$covariate_list)
   if (!is.null(index)) result$index <- index
   result
 }
 
-simulate_many <- function(n_simulations = 1e3, 
-                            formula = status ~ logtitre,
-                            init_seed = NULL,
-                            ...) {
-  future_map_dfr(1:n_simulations, simulate_one, init_seed, formula, ...)
+simulate_many <- function(n_simulations = 1e3, init_seed = NULL, ...) {
+  future_map_dfr(1:n_simulations, simulate_one, init_seed, ...)
 }
 
 summarise_many <- function(simulation_results) {
@@ -75,7 +65,7 @@ summarise_many <- function(simulation_results) {
     ) %>%
     group_by(term, true_value) %>%
     summarise(
-      estimate_mean = mean(estimate), 
+      estimate_mean = mean(estimate),
       estimate_sd = sd(estimate),
       se_mean = mean(std_error),
       coverage = sum(captures_true) / n()
@@ -89,9 +79,18 @@ save_csv <- function(data, folder, name, n_simulations) {
   )
 }
 
-n_simulations <- 10
+cov_list <- list(
+  "logHI" = list(gen_fun = function(n) rnorm(n, 2, 2), true_par = 2),
+  "logNI" = list(gen_fun = function(n) rnorm(n, 2, 2), true_par = 1)
+)
+
+n_simulations <- 1e4
+
 simulation_results <- simulate_many(
-  n_simulations = n_simulations, init_seed = 20191017
+  n_simulations = n_simulations,
+  init_seed = 20191017,
+  covariate_list = cov_list,
+  n = 1e4, lambda = 0.5, beta_0 = -5
 )
 save_csv(simulation_results, this_folder, "result", n_simulations)